Move DataObjectFilters tests
